requirements.txt
requests>=2.31.0
google-generativeai>=0.3.0
openai>=1.0.0
huggingface_hub>=0.19.0
python-dotenv>=1.0.0
